pyproject.toml
[project]
name = "biomind"
version = "1.0.0"
description = "Biomedical research assistant using live arXiv search, BM25 retrieval, Groq LLM, FastAPI, Gradio, and MCP."
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [{ name = "BioMind Team" }]
dependencies = [
  "arxiv==2.1.3",
  "rank-bm25==0.2.2",
  "langchain==0.3.7",
  "langchain-groq==0.2.1",
  "langchain-community==0.3.7",
  "fastapi==0.115.5",
  "uvicorn==0.32.1",
  "python-dotenv==1.0.1",
  "httpx==0.27.2",
  "gradio==4.44.1",
  "mcp==1.0.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = false
